Layout the code for the soft biometrics pipeline
The plan for the pipeline is for 2 soft biometrics: age and gender.
According to Mohit's paper the best performing approaches for the soft
biometrics for age and gender were xgb and RNN respectively. There is
a strange error with the dimensions of the feature vectors for both
algorithms so we need to check that